paged-kv: validate ONNX symbolic path numerically + fix two symbolic bugs
Add an eager-vs-onnxruntime numeric test for the paged ops (the symbolic onnxscript
path is what the AIC compiler consumes; eager parity could not cover it). It caught
two real bugs that only manifest on the ONNX path:
- ScatterND/GatherND index Concat mixed int32 with the int64 head Range -> ONNX
  type error. Fix: keep all paged index tensors int64 (matching position_ids and
  the proven continuous-batching ops), in cache_utils and the ops tests.
- Range used 1-D [0]/[1] args; the ONNX spec requires scalar (0-d) args. Fix: use
  scalar Range bounds (spec-compliant; works on onnxruntime and AIC).

eager vs onnxruntime max_abs_diff = 0.0. Full suite green.

+++ QEfficient/transformers/cache_utils.py
@@ -417,10 +417,13 @@ class QEffPagedDynamicLayer(CacheLayerMixin):
         """
         invalid = position_ids < 0
         logical_block = (position_ids // page_size).clamp(min=0).to(torch.int64)  # [bsz, seq]
-        phys_block = torch.gather(block_table, 1, logical_block).to(torch.int32)
-        offset = (position_ids % page_size).to(torch.int32)
+        # Keep index tensors int64 (matching position_ids) so the onnxscript scatter
+        # builds a single-typed ScatterND index (mixing int32 with the int64 head
+        # Range would break ONNX/onnxruntime type checking).
+        phys_block = torch.gather(block_table, 1, logical_block).to(torch.int64)
+        offset = (position_ids % page_size).to(torch.int64)
         null_block = num_blocks - 1
-        phys_block = torch.where(invalid, torch.tensor(null_block, dtype=torch.int32), phys_block)
+        phys_block = torch.where(invalid, torch.tensor(null_block, dtype=torch.int64), phys_block)
         offset = torch.where(invalid, torch.zeros_like(offset), offset)
         return phys_block, offset
 
@@ -437,9 +440,9 @@ class QEffPagedDynamicLayer(CacheLayerMixin):
         bsz = position_ids.shape[0]
         ctx_indices = torch.arange(ctx_len)[None, :].expand(bsz, ctx_len)  # [bsz, ctx]
         logical_block = (ctx_indices // page_size).to(torch.int64)
-        phys_block = torch.gather(block_table, 1, logical_block).to(torch.int32)
+        phys_block = torch.gather(block_table, 1, logical_block).to(torch.int64)
         phys_block = phys_block.clamp(0, num_blocks - 1)
-        offset = (ctx_indices % page_size).to(torch.int32)
+        offset = (ctx_indices % page_size).to(torch.int64)
         gather_limit = position_ids.max(1, keepdim=True).values  # [bsz, 1]
         invalid_mask = ctx_indices > gather_limit  # [bsz, ctx]
         return phys_block, offset, invalid_mask
